Release 1.1.2 (#24)
* Fix visualization in some methods

* Additional check for invalid face crop
During the inference on the stream from the web camera some of the face box coordinates turned out to be negative. This usually happened when part of the face was out of the frame. This led to invalid crop of the face and error during the inference. Therefore, additional check was added.

* Style fix

* Update version to 1.1.2